Major revisions to get PEACE compiling under Visual Studio for Dr. Ozden. Hopefully this was a worth while effort. The windows code lacks some of the Linux functionality. There are spots in the code where in had to add some unnecessary complexity to get the code to compile with Visual Studio 2008.
The code compiles and links. However, some serious testing is needed to see
if the code actually works as expected under Windows.

Next, a few more changes will be needed to fixed any issues that may have
compiled when working under Linux

+++ peace/src/BatonESTData.cpp
@@ -73,7 +73,7 @@ BatonESTData::getSectionBatons(int nSec) {
     for (int k = 0; k < maxMerValue; k++) {
         if (!batons[k].empty()) {
             sectionBatons[k] = std::vector<Baton>();
-            for (int i = 0; i < batons[k].size(); i++) {
+            for (size_t i = 0; i < batons[k].size(); i++) {
                 int bBeg = batons[k][i].getBegin();
                 if (bBeg >= beg && bBeg < end) {
                     sectionBatons[k].push_back(batons[k][i]);
@@ -88,7 +88,7 @@ void
 BatonESTData::resetBatons() {
     for (int i = 0; i < maxMerValue; i++) {
         if (!batons[i].empty()) {
-            for (int j = 0; j < batons[i].size(); j++) {
+            for (size_t j = 0; j < batons[i].size(); j++) {
                 batons[i][j].sectionCount = -1;
             }
         }
